requirements.txt
Django>=4.2
djangorestframework
django-ninja
pydub>=0.25.1
celery>=5.2
redis>=4.0
rarfile
python-magic
